search

+++ src/js/api/auth/logout.mjs
@@ -1,7 +1,14 @@
 import { remove } from "/src/js/storage/index.mjs";
 
+
 export function setLogOut () {
-    remove.addEventListeners("click", (event) =>{
+
+    const logout = document.querySelector ("#logout")
+
+    if(logout){ 
+        remove.addEventListeners("click", (event) =>{
         localStorage.clear();
+        window.location.reload();
     });
+}   
 }
